fix bugs from email and user

+++ src/main/java/one/wcy/ebookloaningtool/xdz/response/TopUpResponse.java
@@ -4,14 +4,16 @@ import lombok.Getter;
 import lombok.Setter;
 import one.wcy.ebookloaningtool.utils.Response;
 
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
 public class TopUpResponse extends Response {
-    private double balance;
+    private BigDecimal balance;
     private String paymentId;
 
-    public TopUpResponse(String state, String paymentId, double balance) {
+    public TopUpResponse(String state, String paymentId, BigDecimal balance) {
         super(state);
         this.paymentId= paymentId;
         this.balance= balance;
